requirements.txt
anyio==4.12.1
boto3==1.42.68
botocore==1.42.68
certifi==2026.2.25
cramjam==2.11.0
exceptiongroup==1.3.1
fastparquet==2025.12.0
fsspec==2026.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.1.0
numpy==2.2.6
packaging==26.0
pandas==2.3.3
pyarrow==23.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
s3transfer==0.16.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
